Break protocols, enums and such out into their own files
