+++ Composify/AdministrateProjectViewController.swift
@@ -8,21 +8,6 @@
 
 import UIKit
 
-protocol AdministrateProjectDelegate: class {
-    /// User aded a section to a project
-    /// - parameter section: The section that was added
-    func userDidAddSectionToProject(_ section: Section)
-
-    /// The user deleted a section from the project
-    func userDidDeleteSectionFromProject()
-
-    /// The user edited the title of one or more sections and/or the project itself
-    func userDidEditTitleOfObjects()
-
-    /// The user deleted the project
-    func userDidDeleteProject()
-}
-
 class AdministrateProjectViewController: UIViewController {
     // MARK: Properties
