+++ Composify/AudioRecorder/AudioRecorderService.swift
@@ -8,10 +8,6 @@
 
 import Foundation
 
-enum AudioRecorderServiceError: Error {
-    case unableToConfigureRecordingSession
-}
-
 /// An AudioRecorderService can record audio.
 protocol AudioRecorderService {
     /// Initialize the recorder with a URL to the place where the resulting audio should be saved.
