+++ Composify/AudioRecorder/AudioRecorderServiceError.swift
@@ -1,0 +1,13 @@
+//
+//  AudioRecorderServiceError.swift
+//  Composify
+//
+//  Created by Eirik Vale Aase on 16/12/2018.
+//  Copyright © 2018 Eirik Vale Aase. All rights reserved.
+//
+
+import Foundation
+
+enum AudioRecorderServiceError: Error {
+    case unableToConfigureRecordingSession
+}
