+++ Composify/ComposifyObject.swift
@@ -1,0 +1,12 @@
+//
+//  ComposifyObject.swift
+//  Composify
+//
+//  Created by Eirik Vale Aase on 16/12/2018.
+//  Copyright © 2018 Eirik Vale Aase. All rights reserved.
+//
+
+import Foundation
+import RealmSwift
+
+typealias ComposifyObject = Object & FileSystemObject
