+++ Composify/Database/DatabaseObject.swift
@@ -1,0 +1,12 @@
+//
+//  DatabaseObject.swift
+//  Composify
+//
+//  Created by Eirik Vale Aase on 16/12/2018.
+//  Copyright © 2018 Eirik Vale Aase. All rights reserved.
+//
+
+import Foundation
+
+/// The basic object that the common DatabaseService protocol uses.
+protocol DatabaseObject {}
