+++ Composify/Database/DatabaseService.swift
@@ -8,18 +8,6 @@
 
 import Foundation
 
-/// The basic object that the common DatabaseService protocol uses.
-protocol DatabaseObject {}
-
-/// A special object that is implemented by the object that loads
-/// stuff. For example, with the RealmDatabaseService, this is the
-/// ProjectStore object, because it has a reference to all project
-/// identification numbers.
-protocol DatabaseFoundationObject {
-    var identification: String { get }
-    var projectIDs: [String] { get set }
-}
-
 /// The basic interface that a database service must implement.
 /// This way we can abstract away all the bullshit, and just
 /// use these specific type-free methods instead of references
